requirements.txt
Flask==2.0.1
Pillow==8.2.0
Matplotlib==3.4.2
Pygame==2.0.1
transformers==4.9.2
torch==1.9.0
pydub==0.25.1